+++ apps/router-e2e/metro.config.js
@@ -42,13 +42,4 @@ const config = getDefaultConfig(
 // Disable Babel's RC lookup, reducing the config loading in Babel - resulting in faster bootup for transformations
 config.transformer.enableBabelRCLookup = false;
 
-const originalCustomizeFrame = config.symbolicator.customizeFrame;
-config.symbolicator.customizeFrame = (frame) => {
-  const newFrame = originalCustomizeFrame(frame);
-  // This will make frames collapsed the same as in use apps, where expo/packages are in node_modules
-  // and thus collapsed by default.
-  newFrame.collapse ||= new RegExp('expo/packages/.+').test(frame.file);
-  return newFrame;
-};
-
 module.exports = config;